Move class definitions to avoid circular imports
